Accepteret integrationstest
Systemet er integreret så det er funktionelt

+++ GreenHouseMonitor/GreenHouseMonitor/StepperDriver/include/StepperDriver.h
@@ -59,7 +59,7 @@
 int stepperMode; 
 bool isInitialised; 
 bool isWindowOpen;
-int maxNbrRevolutions;
+double maxNbrRevolutions;
 
 
 struct Motor{
